+++ src/NDItk/MultigroupTable/test/MultigroupTable.test.cpp
@@ -136,32 +136,32 @@ void verifyChunk( const MultigroupTable& chunk ) {
   CHECK_THAT( 0.06, WithinRel( chunk.flux().weights()[6] ) );
 
   // reaction cross sections
-  CHECK( false == chunk.crossSections().empty() );
-  CHECK( 2 == chunk.crossSections().numberReactions() );
-  CHECK( 7 == chunk.crossSections().numberGroups() );
-  CHECK( true == chunk.crossSections().hasReaction( 2 ) );
-  CHECK( true == chunk.crossSections().hasReaction( 16 ) );
-  CHECK( false == chunk.crossSections().hasReaction( 102 ) );
-  CHECK( 2 == chunk.crossSections().reactions()[0].reaction() );
-  CHECK_THAT( 0.0, WithinRel( chunk.crossSections().reactions()[0].qvalue() ) );
-  CHECK_THAT( 10.0, WithinRel( chunk.crossSections().reactions()[0].crossSections()[0] ) );
-  CHECK_THAT( 20.0, WithinRel( chunk.crossSections().reactions()[0].crossSections()[1] ) );
-  CHECK_THAT( 30.0, WithinRel( chunk.crossSections().reactions()[0].crossSections()[2] ) );
-  CHECK_THAT( 40.0, WithinRel( chunk.crossSections().reactions()[0].crossSections()[3] ) );
-  CHECK_THAT( 50.0, WithinRel( chunk.crossSections().reactions()[0].crossSections()[4] ) );
-  CHECK_THAT( 60.0, WithinRel( chunk.crossSections().reactions()[0].crossSections()[5] ) );
-  CHECK_THAT( 70.0, WithinRel( chunk.crossSections().reactions()[0].crossSections()[6] ) );
-  CHECK( 16 == chunk.crossSections().reactions()[1].reaction() );
-  CHECK_THAT( 1.1234567, WithinRel( chunk.crossSections().reactions()[1].qvalue() ) );
-  CHECK_THAT( 1.0, WithinRel( chunk.crossSections().reactions()[1].crossSections()[0] ) );
-  CHECK_THAT( 2.0, WithinRel( chunk.crossSections().reactions()[1].crossSections()[1] ) );
-  CHECK_THAT( 3.0, WithinRel( chunk.crossSections().reactions()[1].crossSections()[2] ) );
-  CHECK_THAT( 4.0, WithinRel( chunk.crossSections().reactions()[1].crossSections()[3] ) );
-  CHECK_THAT( 5.0, WithinRel( chunk.crossSections().reactions()[1].crossSections()[4] ) );
-  CHECK_THAT( 6.0, WithinRel( chunk.crossSections().reactions()[1].crossSections()[5] ) );
-  CHECK_THAT( 7.0, WithinRel( chunk.crossSections().reactions()[1].crossSections()[6] ) );
-  auto xs = chunk.crossSections().crossSection( 2 );
-  CHECK( 2 == xs.reaction() );
+  CHECK( false == chunk.reactionCrossSections().empty() );
+  CHECK( 2 == chunk.reactionCrossSections().numberReactions() );
+  CHECK( 7 == chunk.reactionCrossSections().numberGroups() );
+  CHECK( true == chunk.reactionCrossSections().hasReaction( 2 ) );
+  CHECK( true == chunk.reactionCrossSections().hasReaction( 16 ) );
+  CHECK( false == chunk.reactionCrossSections().hasReaction( 102 ) );
+  CHECK( 2 == chunk.reactionCrossSections().reactions()[0].identifier() );
+  CHECK_THAT( 0.0, WithinRel( chunk.reactionCrossSections().reactions()[0].qvalue() ) );
+  CHECK_THAT( 10.0, WithinRel( chunk.reactionCrossSections().reactions()[0].crossSections()[0] ) );
+  CHECK_THAT( 20.0, WithinRel( chunk.reactionCrossSections().reactions()[0].crossSections()[1] ) );
+  CHECK_THAT( 30.0, WithinRel( chunk.reactionCrossSections().reactions()[0].crossSections()[2] ) );
+  CHECK_THAT( 40.0, WithinRel( chunk.reactionCrossSections().reactions()[0].crossSections()[3] ) );
+  CHECK_THAT( 50.0, WithinRel( chunk.reactionCrossSections().reactions()[0].crossSections()[4] ) );
+  CHECK_THAT( 60.0, WithinRel( chunk.reactionCrossSections().reactions()[0].crossSections()[5] ) );
+  CHECK_THAT( 70.0, WithinRel( chunk.reactionCrossSections().reactions()[0].crossSections()[6] ) );
+  CHECK( 16 == chunk.reactionCrossSections().reactions()[1].identifier() );
+  CHECK_THAT( 1.1234567, WithinRel( chunk.reactionCrossSections().reactions()[1].qvalue() ) );
+  CHECK_THAT( 1.0, WithinRel( chunk.reactionCrossSections().reactions()[1].crossSections()[0] ) );
+  CHECK_THAT( 2.0, WithinRel( chunk.reactionCrossSections().reactions()[1].crossSections()[1] ) );
+  CHECK_THAT( 3.0, WithinRel( chunk.reactionCrossSections().reactions()[1].crossSections()[2] ) );
+  CHECK_THAT( 4.0, WithinRel( chunk.reactionCrossSections().reactions()[1].crossSections()[3] ) );
+  CHECK_THAT( 5.0, WithinRel( chunk.reactionCrossSections().reactions()[1].crossSections()[4] ) );
+  CHECK_THAT( 6.0, WithinRel( chunk.reactionCrossSections().reactions()[1].crossSections()[5] ) );
+  CHECK_THAT( 7.0, WithinRel( chunk.reactionCrossSections().reactions()[1].crossSections()[6] ) );
+  auto xs = chunk.reactionCrossSections().reaction( 2 );
+  CHECK( 2 == xs.identifier() );
   CHECK_THAT( 0.0, WithinRel( xs.qvalue() ) );
   CHECK_THAT( 10.0, WithinRel( xs.crossSections()[0] ) );
   CHECK_THAT( 20.0, WithinRel( xs.crossSections()[1] ) );
@@ -170,8 +170,8 @@ void verifyChunk( const MultigroupTable& chunk ) {
   CHECK_THAT( 50.0, WithinRel( xs.crossSections()[4] ) );
   CHECK_THAT( 60.0, WithinRel( xs.crossSections()[5] ) );
   CHECK_THAT( 70.0, WithinRel( xs.crossSections()[6] ) );
-  xs = chunk.crossSections().crossSection( 16 );
-  CHECK( 16 == xs.reaction() );
+  xs = chunk.reactionCrossSections().reaction( 16 );
+  CHECK( 16 == xs.identifier() );
   CHECK_THAT( 1.1234567, WithinRel( xs.qvalue() ) );
   CHECK_THAT( 1.0, WithinRel( xs.crossSections()[0] ) );
   CHECK_THAT( 2.0, WithinRel( xs.crossSections()[1] ) );
